libinput: Upgrade 0.21.0 -> 1.1.4
This is backport from oe-core/master, commit 7f6a5d2.

yocto/jethro provides the version 0.21 from August 2015 while the
current version is 1.1.4.

The patch libinput-configure.ac-add-arg-with-libunwind.patch has been
merged to upstream and was removed from the recipe. The patch
touchpad-serial-synaptics-need-to-fake-new-touches-on-TRIPLETAP.patch
has not been merged to upstream and is still included in the recipe.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>